requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==2.0.0
pybit==5.7.0
pycryptodome==3.20.0
requests==2.32.3
talipp==2.3.1
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
